major update, add game-agnostic online comm menu
This adds the online comm menu that's used to communicate with the matchmaking server.

It's designed to be mostly game agnostic. Each game will need to define memory addresses for the menu to send callbacks for when the comm menu is exited, or when a match has been found and it's ready to start the battle.