major update, remove old connection menu, use new comm menu
Thanks to XKirby for adjusting the old connection initialization code to make use of our new matchmaking api. The matchmaking api was created by jamesking, along with its own set of server-side code (not pictured here).
There's plenty of other smaller changes too. This file is still largely in the debugging phase.
The old method of connecting is still available, it's been renamed to "Direct IP" and is accessible from within the Online Comm Menu.